refactor: split App.tsx round 2 (5 hooks) and start ipc handlers domain split
App.tsx (4598 -> 3886 lines, -15%):
Continue the round 1 effort by extracting 5 more hooks from
AppMainWorkspaceInner. AppMainWorkspaceInner now has another ~700 lines
of self-contained logic moved out, with no behavioral changes (every
useEffect order, IPC timing and ref initialization preserved).

New hooks:
- useTeamSessionActions: select team task / approve / reject plan proposal
- useAgentSessionActions: 7 agent sub-session callbacks (select / send / wait / resume / close / etc.)
- useEditorCenterDerived: 10 derived state pieces for the editor center pane (monaco doc path, markdown view, plan chrome, ...)
- usePlanWizardActions: plan question + user input + plan build + plan execute + skill creator flow
- useAgentSidebarFilePreview: 370-line openAgentSidebarFilePreview with its 4-source diff fallback

(useShellLayoutProps was considered but skipped: each of the 5 layout-props
useMemos already has 60+ deps, so externalizing yields ~0 net line savings.)

main-src/ipc/register.ts (3917 -> 3783 lines):
Begin domain-based split of the 100+ ipcMain.handle file. First batch
covers the 4 most independent domains (no file-private helper deps):
- handlers/clipboardHandlers.ts: clipboard:writeText / readText
- handlers/autoUpdateHandlers.ts: 4 auto-update commands
- handlers/usageStatsHandlers.ts: usageStats:get / pickDirectory
- handlers/lspHandlers.ts: 4 ts LSP commands

senderWorkspaceRoot is now exported so handler files can reuse the same
workspace-root resolution as the in-file handlers. registerIpc() now
calls register*Handlers() for each split-out domain. Other domains
(git / agent / threads / fs / shell / plugins / mcp / browser / settings /
app / workspace) depend on file-private helpers (runChatStream and ~30
other helpers) and will be split in a follow-up pass.

Verified after each phase with `npx tsc --noEmit` and `npm test`
(377/377 passing). No linter errors in any new file.

Made-with: Cursor
